+++ sscRaft/filters/phase_filters.py
@@ -32,13 +32,13 @@ def phase_filters(tomo,dic):
         energy    = dic['energy[KeV]']
 
 
-    alpha     = dic['regularization']
+    alpha     = dic['phase filter regularization']
     pixel     = dic['detectorPixel[m]']
 
     padx = int(nrays // 2)
     pady = int(nrays // 2)
     try:
-        pad  = dic['phase pad']
+        pad  = dic['padding']
         padx = int(pad * padx)
         pady = int(pad * pady)
         # padx = power_of_2_padding(nrays,padx)
@@ -74,16 +74,26 @@ def phase_filters(tomo,dic):
         logger.error(f'Data has wrong shape = {tomogram.shape}. It needs to be 2D or 3D.')
         sys.exit(1)
 
-    if nangles == 1:
-        blocksize = 1
-    else:
+    try:
         blocksize = dic['blocksize']
+        
+        if nangles == 1:
+            blocksize = 1
+        else:
+            blocksize = dic['blocksize']
+    except:
+        blocksize = 32
+
+        if nangles == 1:
+            blocksize = 1
+        else:
+            blocksize = 32
 
     if blocksize > ( nangles // ngpus ):
         logger.error(f'Blocksize is bigger than the number of angles ({nangles}) divided by the number of GPUs selected ({ngpus})!')
         sys.exit(1)
 
-    filtername = dic['phase filter']
+    filtername = dic['phase filter method']
     filter     = PhaseFilterNumber(dic['phase filter'])
     
     if filter == 0:
@@ -91,7 +101,7 @@ def phase_filters(tomo,dic):
         logger.warning(f'Finishing Run ...')
         sys.exit(0)
     else:
-        logger.info(f'Phase filter: {filtername}({filter})')
+        logger.info(f'Phase filter method: {filtername}({filter})')
 
     float_param     = numpy.array([z1x, z1y, z2x, z2y, energy, alpha])
     float_param     = numpy.ascontiguousarray(float_param.astype(numpy.float32))